+++ libs/adapter/event-store-db/event-store-db.event-subscription.spec.ts
@@ -3,7 +3,7 @@ import { EventListener } from '@pl-oss/domain';
 import { EventStoreDBEventSubscription } from './event-store-db.event-subscription';
 
 class TestEventListener extends EventListener {
-  getStreamPrefixes(): string[] { return []; }
+  getStreamNamePrefixes(): string[] { return []; }
 }
 
 async function* generateEvents() {
@@ -34,7 +34,7 @@ describe('EventStoreDBEventSubscription', () => {
 
   describe('register', () => {
     it('should register and call listener.on', async () => {
-      const prefixes = testEventListener.getStreamPrefixes();
+      const prefixes = testEventListener.getStreamNamePrefixes();
       const filter = streamNameFilter({ prefixes });
       const events = generateEvents();