+++ libs/domain/test-domain/test.aggregate.ts
@@ -2,7 +2,7 @@ import { Aggregate } from '../lib/aggregate';
 import { TestEvent } from './test.event';
 
 export class TestAggregate extends Aggregate {
-  getStreamPrefix(): string { return 'TestAggregate'; }
+  getStreamNamePrefix(): string { return 'TestAggregate'; }
 
   // eslint-disable-next-line
   applyTestEvent(testEvent: TestEvent) {}