+++ libs/domain/test-domain/test.event-listener.ts
@@ -1,7 +1,7 @@
 import { EventListener } from '../lib/event-listener';
 
 export class TestEventListener extends EventListener {
-  getStreamPrefixes(): string[] { return []; }
+  getStreamNamePrefixes(): string[] { return []; }
 
   // eslint-disable-next-line class-methods-use-this
   async onTestEvent(): Promise<void> { return Promise.resolve(); }